RYSK-300: Make logic reusable.
