+++ packages/front-end/src/components/optionsTrading/Modals/AdjustCollateralModal/index.tsx
@@ -13,6 +13,7 @@ import { Button, Input, Label, Wrapper } from "../Shared/components/Form";
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
+import { roundInputValue } from "../Shared/utils/roundNumberValue";
 import { Pricing } from "./components/Pricing";
 import { Symbol } from "./components/Symbol";
 import { Toggle } from "./components/Toggle";
@@ -39,12 +40,7 @@ export const AdjustCollateralModal = () => {
     useNotifications();
 
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const amount = event.currentTarget.value;
-    const decimals = amount.split(".");
-    const rounded =
-      decimals.length > 1
-        ? `${decimals[0]}.${decimals[1].slice(0, 2)}`
-        : event.currentTarget.value;
+    const rounded = roundInputValue(event);
 
     setAmountToAdjust(rounded);
   };
