+++ packages/front-end/src/components/optionsTrading/Modals/CloseOptionModal/index.tsx
@@ -24,6 +24,7 @@ import {
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
+import { roundInputValue } from "../Shared/utils/roundNumberValue";
 import { Pricing } from "./components/Pricing";
 import { usePositionData } from "./hooks/usePositionData";
 
@@ -57,20 +58,14 @@ export const CloseOptionModal = () => {
   };
 
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const amount = event.currentTarget.value;
-    const decimals = amount.split(".");
-
-    const rounded =
-      decimals.length > 1
-        ? `${decimals[0]}.${decimals[1].slice(0, 2)}`
-        : event.currentTarget.value;
+    const rounded = roundInputValue(event);
 
     const maxAmount = Math.min(
       positionData.totalSize,
       parseFloat(rounded || "0")
     );
 
-    setAmountToClose(maxAmount ? maxAmount.toString() : amount);
+    setAmountToClose((maxAmount ? maxAmount : rounded).toString());
   };
 
   const handleApprove = async () => {
