+++ packages/front-end/src/components/optionsTrading/Modals/CloseShortOptionModal/index.tsx
@@ -26,6 +26,7 @@ import {
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
+import { roundInputValue } from "../Shared/utils/roundNumberValue";
 import { Pricing } from "./components/Pricing";
 import { useShortPositionData } from "./hooks/useShortPositionData";
 
@@ -59,13 +60,7 @@ export const CloseShortOptionModal = () => {
   };
 
   const handleChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const amount = event.currentTarget.value;
-    const decimals = amount.split(".");
-
-    const rounded =
-      decimals.length > 1
-        ? `${decimals[0]}.${decimals[1].slice(0, 2)}`
-        : event.currentTarget.value;
+    const rounded = roundInputValue(event);
 
     setAmountToSell(rounded);
   };
