+++ packages/front-end/src/components/optionsTrading/Modals/SellOptionModal/index.tsx
@@ -15,6 +15,7 @@ import { Button, Input, Label, Wrapper } from "../Shared/components/Form";
 import { Header } from "../Shared/components/Header";
 import { Modal } from "../Shared/components/Modal";
 import { getButtonProps } from "../Shared/utils/getButtonProps";
+import { roundInputValue } from "../Shared/utils/roundNumberValue";
 import { Filters } from "./components/Filters";
 import { Pricing } from "./components/Pricing";
 import { Symbol } from "./components/Symbol";
@@ -42,12 +43,7 @@ export const SellOptionModal = () => {
     useNotifications();
 
   const handleAmountChange = (event: ChangeEvent<HTMLInputElement>) => {
-    const amount = event.currentTarget.value;
-    const decimals = amount.split(".");
-    const rounded =
-      decimals.length > 1
-        ? `${decimals[0]}.${decimals[1].slice(0, 2)}`
-        : event.currentTarget.value;
+    const rounded = roundInputValue(event);
 
     setAmountToSell(rounded);
   };
